Fixed 1.9 parser GC issue.
References creating during parsing were not in memory that the MRI GC
could not see. If the process ran long enough, the references would
get collected and the process would crash.

In rbx, the C-API handles prevented the objects from getting collected
because the method that called into C was still on the stack, and the
handles are associated with that native method frame.

The solution is to keep a list of references created during parsing and
root that list in the Melbourne instance used to process the C-struct
parse tree into an AST of Ruby objects.