Store quarks in rb_parser_state
Issue #1308 is actually due to a race condition in the
parser. The problem was that the quarks were stored globally
so if multiple threads are compiling something they would
both modify this global list.

This rework ensures we store the quarks in the rb_parser_state
which isn't global so multiple compiles simultaneous won't
cause crashes like #1308.

Fixes #1308