made alloca.h includes only for solaris
In September alloca.h was added to build on solaris systems. The problem
there was, that alloca.h was not included through stdlib.h.

FreeBSD has no alloca.h which can be included, because everything is in
stdlib.h and therefore rubinius could not be built.

On Linux, it's no problem, as alloca.h gets included through stdlib.h.

The fix is to check in configure, if alloca.h can be included and set a
constant to check in the source files if alloca.h can be included or
not.

The constant is named HAVE_HEADER_ALLOCA_H as HAVE_ALLOCA and
HAVE_ALLOCA_H are already used and set by ffi.